Update references from flattening out `src/sandbox/`
