+++ src/sandbox/startup-scripts.js
@@ -1,6 +1,6 @@
 let { exec } = require('child_process')
 let series = require('run-series')
-let { userEnvVars } = require('../../lib')
+let { userEnvVars } = require('../lib')
 
 module.exports = function startupScripts (params, callback) {
   let { cwd, inventory, update, restart, runStartupCommands = true } = params
